apps/api/pyproject.toml
[project]
name = "citetrack-api"
version = "0.0.0"
description = "Citetrack AI — FastAPI backend for AI visibility tracking"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.1",
    "loguru>=0.7.2",
    "httpx>=0.28.0",
    "solaraai-llm[agent]==0.0.35",
    "mirascope<2",
    "solaraai-prisma-client",
    "arq>=0.27.0",
    "fpdf2",
    "svix>=1.40.0",
]

[tool.uv.sources]
solaraai-llm = { path = "/home/yuval/Documents/solaraai/stanley/repos/solaraai-packages/packages/python/solaraai-llm", editable = true }
solaraai-prisma-client = { path = "/home/yuval/Documents/solaraai/stanley/repos/prisma/dist/client-python", editable = true }

[dependency-groups]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "pytest-playwright",
    "tomlkit",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: tests that call real external APIs (OpenAI, Tavily)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "ASYNC",  # flake8-async
]
preview = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["ASYNC"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false


